Fix the cephosd server reimages
A recent reimage of cephosd1001 failed because the custom ceph_disks
fact is not present on the host until perccli64 is available and the
ceph-mon package is installed. This caused a compilation error on the
first puppet run.

In order to address this we can simply put a conditional around the
block that tries to use the fact to populate the $storage_disks
variable.

This will result in the Ceph OSDs not being managed during the first
puppet run, which is not an ideal solution. However, we may be moving
away from puppet managed OSDs anyway, so we can view this as a temporary
solution to fix the reimage process.

Hosts: O:ceph::server
Bug: T362993
Change-Id: Ia1c6c59df2f53d61c7ffecbeee04e587678bceb6