phabricator::logmail: parameterize sender adddress for stats mails
Instead of repeating the same string over and over again let's
move it to Hiera where data like this belongs more than in the
middle of puppet code.

Also easier to change if we ever want to change it.

We could also do this with the various recipient addresses and
have it all in one place if we can find names for "recipient sets".

Change-Id: I030a4aa05d8bdf53c23c16f2a66bd27f53e6d8af