Fix the ceph osd activate command for hdd devices
During recent reimages we have noticed that the ceph commands to actvate
a new OSD fail when the device is a hard drive with a separate bluestore
WAL device. This is do to the way that we obtain the osd_fsid for the
new device.

This commit adds an alternative operator that removes the 'null' value
returned by the osd_fsid with an empty string.

Bug: T362993
Change-Id: I340027bd87a4715be8cabc317cf32ff07e481b1a