Allow the ceph-common package to create the ceph user/group
We have previously been creating the ceph user and group prior to
installing the ceph-common package. The result is that the system user
and group is allocated automatically from the list of allowed uid/gid
values. However, it may change between one installation and the next,
which can present a problem when reimaging servers.

The preferred mechanism is to allow the ceph-common package to create
the system user and group itself and this will default to the
pre-allocated value of 64045.

See: https://github.com/ceph/ceph/blob/main/debian/ceph-common.postinst#L31

This change causes the ceph group to become dependent on the
installation of the ceph-common package, rather than the other way
around.

This will help when we want to start reimaging servers and keeping the
contents of the /var/lib/ceph directory between reimages.

For completeness, we also explicitly cause the file resources to depend
on the creation of the user account, rather than depend on the order in
the file.

Hosts: O:ceph::server,O:wmcs::ceph::mon,O:wmcs::ceph::osd
Bug: T362993
Change-Id: Ic21670087cca74f7f46d22cb1b246ef6eb44b98d