magru: set hiera for trafficserver::backend::storage_elements
The real reason for setting this right now and not in a later commit is
to create the directory hieradata/magru as otherwise, the lookup()
against hieradata/magru.yaml fails. See T362729#9756089 for a more
detailed explanation. But since we need it later anyway, let's set it
now to unblock all reimaging.

Bug: T362729
Bug: T346722
Change-Id: I77ac20c7ce5e5c57c0a465761e2eb39a25d8bc32