mariadb: Upgrade db1150, db1171 and move s4, s7, s8 backups to 10.6
db1150, db1171 will be upgraded to MariaDB 10.6.
Backups (snapshots) for s4, s7 and s8 will start happening in
MariaDB 10.6 format.

dbprov1002/3 and dbprov2002/3 will only be left with misc backups,
easyly movable to the other hosts.
dbprov1001 and dbprov2001 will only be left with s1, the only
pending section to be upgraded to 10.6.

db2098 will be decomissable.

Bug: T362509
Bug: T360751
Change-Id: I86f62c14acd06cef83ac64b64fc6218930e5a6df