Test a fix for the bootstrapping of mon daemons on cephosd*
When reimaging the cephosd servers the initial bootstrap of the mon
daemons fails and requires manual intervention. The reason seems to be
that the temporary keyring that is written to
/var/lib/ceph/tmp/client.mon.keyring contains the hostname of the mon
server, where it should contain a wildcard 'mon.' that refers to all mon
servers.

This commit causes the ceph cluster running on cephosd100[1-5] to write
a new keyring with this updated format, so that we can test
bootstrapping monitor daemons.

It is intended to cause no changes to the 'cloudceph' clusters that are
operated by the WMCS team, since these are in production use. We will
review at a later date whether or not the fix is also appropriate for
those clusters.

Hosts: O:ceph::server,O:wmcs::ceph::mon,O:wmcs::ceph::osd
Bug: T332987
Bug: T362993
Change-Id: I8696a1fb8ddd7e9c17dc87a05532cdb3ffe578c1