mediawiki/geoip: make loading geoip data from puppetserver optional
We don't need geoip data for deployment servers in cloud.

Just deployment_server includes mediawiki::commmon in production
which in turn includes geoip as "needed for MediaWiki".

This results in puppet errors if you have a puppetserver
and a deployment server in cloud VPS.

Make it optionable to fetch the geoip data from puppet so it
can be disabled in Hiera.

Bug: T363415
Bug: T360964
Bug: T370436
Change-Id: I5f9b7aaa58b59b43671bba56b4faa8529671652d