memcached: add memcached_user option
All our memcached services run under the 'nobody' user, which
is also hardcoded memcached.systemd.erb.

In this patch we are adding the option of defining which
user memcached will run as. Defaulting to 'undef' for now,
so we can migrate each service one by one.

After having all services running as the 'memcache' user, we
can switch this default accordinly.

Bug: T273950
Change-Id: Iff99abc3d832c83b1f81396e5a88783653c0820b