Add the wmf-java-cacerts truststore to all remaining airflow hosts
We are about to enable TLS by default for connections from Airflow DAGs
to the aqs* cassandra cluster. As such we will need to ensure that the
Java truststore that contains our pki root certificate is installed to
all airflow nodes.

It is already present on an-launcher, where the analytics instance of
Airflow currently runs, but this commit adds it to the remaining
instance hosts.

Hosts: P:airflow
Bug: T362181
Change-Id: I88577e665deedf369ec474b735adadd0b00d5ec3