maintain-views: Redact entire logging_logindex rows without a target
The logging_logindex table is supposed to be used for the indexes on
log_page or (log_namespace, log_title). For this to be effective, rows
where those are NULL must be filtered out entirely instead of those
columns to be redacted but the row to be visible.

Follows-up Id19f88ce7a331e3f2f55c4e84d2cc295bc255179 which did this the
wrong way and made the views much slower than necessary.

Bug: T363633
Bug: T364151
Change-Id: Icc2fc2d73542c73df484c01017d0954d03e33783