confd: Extend confd-lint-wrap to accept a unique resource name
General idea: service_template.toml.erb will be updated to pass the
unique file-safe identifier for this resource to the wrapper, which
is in turn used to name the error state file.

This makes it so that confd_prometheus_metrics.py no longer needs to
reconstruct the error state file name by reverse engineering the confd
staged file name format.

This also has the benefit that a given confd-managed file will only ever
have one error state file (rather than having a random suffix), which is
more in line with the intent of confd-lint-wrap (see cleanup logic in
the success case).

Bug: T363924
Change-Id: I859e13d7801e19463ef111227b9b9c7f958dc03a