Move stats misc_jobs from stat1007 to stat1011
The stat1007 server currently has a number of miscellaneous jobs running
for the wmde and product_analytics teams.

Since we plan to decommission stat1007 in the near future, we need to
find a new home for these jobs, so this commit makes stat1011 that new
home.

In the longer term we will be moving all of these jobs to airflow, but
we still need them to run on a stat host for the time being.

Hosts: stat1007.eqiad.wmnet,stat1011.eqiad.wmnet,clouddumps1002.wikimedia.org
Bug: T353785
Change-Id: I7ebfa46cbc18cff64cae374853fdca5c9d8c460f