confd: confd-lint-wrap ignores positional args separator
This is part one of two patches that prepare confd-lint-wrap for
adopting argparse.

Since the script accepts a check command that potentially contains
command-line flags, a positional argument separator will be used to
avoid confusing argparse (and needs to happen _first_).

Introducing a separator is a two-step process:
1. teach confd-lint-wrap to ignore the separator if present (this)
2. add the separator to check_cmd in the puppet template (next)

Once that's done, we can switch over to using argparse (next next).

Bug: T363924
Change-Id: Ib76de3def7ef62c35a77dcfa53d2faadc470e7e8