Drop the deprecated dumps fetcher that pulls from stat1007
We are in the process of decommisisoning stat1007, so we need to stop
any jobs that regularly pull data from this host to the dumps web
servers.

There were three jobs that ran on this host:
* wikistats_1
* pagecounts-ez
* media-contestwinners.

The files are synced and publicly available at

* https://dumps.wikimedia.org/other/wikistats_1/
* https://dumps.wikimedia.org/other/pagecounts-ez/
* https://dumps.wikimedia.org/other/media/contest_winners/

However, the source files on stat007 have not been updated since
September 2020 so I don't believe that these fetcher jobs need to be
migrated to a new stat host.

Hosts: O:dumps::distribution::server
Bug: T353785
Bug: T238243
Change-Id: I3228368cc142f2e95ade27e5dec9a5e377df91c1