Move dumps::generation::worker::dumper_misc_crons_only role
We are preparing to decommission snapshot1008 so we need to move the
role that carries out the miscellaneous dumps from this host to a new
host. The new host selected for the role is snapshot1017.

We can also remove the reference to snapshot1009 from site.pp as this
host has now been decommissioned.

Hosts: snapshot1008.eqiad.wmnet,snapshot1017.eqiad.wmnet
Bug: T325228
Change-Id: I4079f2c874cf1b864ab8163cc19493c069fa4092