ci: include envoyproxy in ci_test role
The production ci role and the ci_test
role are similar but not identical.

Both include profile::ci::website and profile::ci::httpd
but envoyproxy was not included in ci_test so far.

Recently in Iae3eec3f0506 auto_restart::service for envoyproxy
was added to profile::ci::website.

This errors out when it tries to restart envoyproxy which doesn't
exist.

The easiest fix seems to just included envoyproxy here as well and
reduce the difference between both roles.

Also the main purpose of this machine was testing zuul which has
already been done. To be determined if this stays permanently now.

For now just want to fix T364510 and there isn't much to break on
this machine.

Bug: T364510
Change-Id: Icddb2d65de9a553b41bb48c1685bb2f2f80b325d