cache:benthos: test for socket based activation in Benthos
With this test we're going to create a systemd activated socket that
can be tuned to an arbitrary ReceiveBuffer value. The new socket
activates the corresponding Benthos instance (a new one) and pass the
newly created socket as stdin.

This is controlled by profile::cache::base::benthos_systemd_socket to
activate the socket unit and relative Benthos unit override.

This is mutually exclusive with
I3ee130e55ba68f5ff8e80d67b33a419b9ab57b58 and should not be merged
together.

Bug: T364379
Change-Id: I06f17a7225d75862000f9a0a6423ae499de4309a