Add a class to Cumin hosts which generates a Kafka certificate for frtech
The fr-tech environment is separate from production, but the Kafka cert
is currently managed in cergen and manually deployed within fr-tech.

Since we're deprecating cergen, issue the cert from the PKI instead.
The cerficate is installed on the cumin hosts since fr-tech-admins
already have access to these servers.

Those PKI-issued certs have a validity of 180 days, when the procedure
is generally known to work, we can add a systemd timer which checks
the remaining cert validity and gives a advance warning when the expiry
is upcoming.

Hosts: cumin1002.eqiad.wmnet
Bug: T360779
Change-Id: I419d0d77cf5f34461c398b3784f7820aa1b01dcf