varnish: Copy value of X-Wikimedia-Debug cookie to header
To facilitate debugging on mobile devices, treat requests with an
X-Wikimedia-Debug cookie identically to requests with an
X-Wikimedia-Debug header. To minimize complexity, do this at the
edge by having Varnish copy the cookie value to the header.

Varnish is not well equipped for cookie parsing, but this cookie
is generated by us in PHP, so we can get away with a minimum amount
of decoding. Attempts to set the cookie manually will probably not
work.

To prevent people accidentally forgetting the debug flag on, the
cookie has an expiry parameter and a cookie with an expiry value
that's either in the past or too far in the future is ignored.
Replicating this logic in Varnish seems too hard to be worth it,
but PHP will automatically remove ignored cookies, so not doing
that here shouldn't be a problem.

Bug: T350094
Change-Id: Id292703c6f76a9587d4cbfcfd807893afd9f3ac0