Improve dumps::web::rsync::nginxlogs management
We observed an orphan systemd timer operating on the standby clouddumps
web server, which had been responsible for pushing the
dumps.wikimedia.org web logs to a stat server for analysis.

This change adopts that orphaned service and deletes it on the standby
server.

Hosts: clouddumps1001.wikimedia.org,clouddumps1002.wikimedia.org
Bug: T364820
Change-Id: Ic1b20c860f6dfe0b190e5975aa830c7b3eb79e6e