Move Swift on thanos-fe1001 to PKI TLS cert
Force ms-fe1001 to request a PKI TLS cert for its envoy
instance. We want to rollout the change to a single node
since in the past we tried the whole cluster and Tegola (maps)
showed some slowness and increase in CPU usage (ending up
in an outage).

The goal is to test again Tegola with a single Thanos Swift
node upgraded, and decide what to do after reviewing Tegola's
metrics for a few days. In case of fire, we can easily depool
thanos-fe1001 and revert, limiting the impact to Tegola's clients.

The hosts set in the cfssl_options are the SANs gatheted
from the current TLS cert's list, minus all the ones already
served by titan nodes.

Bug: T344324
Bug: T356412
Change-Id: Id412a7ed050fe56c1b7a16ac6b31139f5dc5ec5a