Clarify totoro.wikimedia.org test
This test randomly failed during a scap deployment, and there was much
confusion in #wikimedia-operations: nobody had heard of this host name,
there is no trace of it in operations/dns.git, and the change which
added it and other httpbb tests (change I70e5ce39f4, commit ee67f8d5ba)
includes no information about this particular host name. I can only
assume that it’s meant to be an arbitrary subdomain that doesn’t need to
exist in DNS (since the test bypasses DNS); add a comment to clarify
this and rename the host intentionally-no-such-domain.wikimedia.org,
which is hopefully more self-explaining.

Bug: T364880
Change-Id: I5e8802c0f4db48155d6dbcdca038b934c50e4d88