stewards: add rsync server, let lists primary host pull data
Add an rsync server to the steward hosts.

Create an rsync server module that lets the lists primary host
pull stewards subscriber data from the export directory.

(currently: lists1001.wikimedia.org from /srv/exports)

The synced files will then be used to automate list subscriptions
for stewards.

To be able to find the name of the lists primary host we have to
add it to common.yaml because otherwise it's only applied
to hosts with the list server role via role/common/.

We could later switch the lists role itself to also use that value
so we don't have it in 2 places.

Bug: T351202
Change-Id: Iac64e2b0acc7bb9619f31c9ef415674450db5968