pontoon: add frontend proxy capability to LB
Cloud VPS webproxy can be configured to proxy a whole zone e.g.
'*.$public_domain' to a backend instance.

Tweak the LB to be able to receive such requests (over http) and
HTTP-proxy them to the service backend.

Previously, the same was achieved with an additional instance running
pontoon::frontend role and consuming one floating IP.

Bug: T352640
Change-Id: I5089e51c9a0807ef6091cd0755f2473935471f61