ci/zuul: use localhost as gearman server
follow-up to Ieabd362b6689a8

We now use a DNS lookup rather than an IP in Hiera
to avoid having to make manual changes.

There are separate configs for zuul-server and zuul-merger.

If we keep using 127.0.0.1 here:

DNS lookup failed for 127.0.0.1 because it looks for an A record
rather than PTR.

Bug: T334517
Change-Id: Ia85e9dae3cb11e5427cd9e4b061c4139be04b84e