lists: move definition of primary and standby host to common hiera
To be able to lookup the primary list server in other profiles
not using the same role.

Like from the stewards hosts to allow the primary server to rsync.
(follow-up to Iac64e2b0acc7bb)

Following the same pattern we have for a lot of other "primary host"
definitions in Hiera where we solved the same issue by moving them
to a common place.

Change-Id: Ie5e5a7c832cf6892816b5f4663ab57661cbac999