gerrit: fix team name for https monitor alerting
The team name changed from "sre-collab" to "collaboration-services".

With it the special receiver for the combined teams of collaboration
services and releng was renamed accordingly for all checks but we
missed this one place.

This meant when the alert got triggered recently it reported only
on IRC in -operations but did NOT also:

- report in collaboration services channel
- report also in releng channel
- also send email to the releng team list
- also create a Phabricator ticket.

After this fix it should do all these things again.

Bug: T365148
Bug: T329587
Change-Id: I1d7874f53126fc3433cf03e21a6aef63cdb17ebf