scap: remove snapshot1008 from dsh group mediawiki-installation
This host is being decom'ed but wasn't removed from scap "dsh" groups
lists yet.

The "mediawiki-installation" group consists of a bunch of conftool
groups AND a list of host names.

snapshot hosts are special since they are not in conftool but ALSO get
mediawik installations.

This lead to errors for mediawiki deployers when running scap.

Bug: T325228
Change-Id: Ibe40e2070601688b46007414a01fe16c03ed9aca