Pass Druid middle manager ports as port range
Ferm simply gets a string and then parses it, so the value read from
Hiera for the Peon ports gets accepted. When using nftables however
we've introduced stricter data types which allow actual type validation.

As such, pass the Peon ports as a $port_range which is compatible with
both ferm and nftables.

Hosts: C:profile::druid::middlemanager
Change-Id: I2db27b026dcee8d65908457fbeee4db7bba76e2b