Run Gobblin later to let time for Canary events
Now that we run the generation of Canary events from Airflow, we need
to run it twice an hour to make sure Gobblin generates the _IMPORTED
flag.

In order to let some time for Airflow to generate the canary events at
the beginning of the hour (HH:02), we should run Gobblin slightly later.

Bug: T365223
Change-Id: I3286efe7595611b8c6aa153aea83fb54baa0a496