Skip ROCm packages for ml-staging2001
The host is going to be reimaged to Bookworm, and the ML
team is going to test if the ROCm packages are needed
on the bare metal k8s node. The current assumption is that
the libraries/drivers/etc.. will only be needed on
the container itself, the k8s node will only need to load
drivers shipped with the Kernel.

Bug: T363191
Change-Id: I9e4d3d08a0af671d1231a3ea398c546f8a22cd14