P:idm Use account login page for monitoring.
Signups have been disabled for CloudIDM, causing monitoring to fail.
Instead use the /accounts/login page, as this is available on all Bitu
installations. The login form does not work on production IDM, but is
rendered by the Django framework, ensuring that we test the uWSGI
process and not just Apache.

Change-Id: I3b8bf551fcc545ed7f18309d8a436298c3796a53