lists: add timer to sync data from stewards hosts
Iac64e2b0acc7bb added an rsync server on stewards hosts.

Now we can pull subscriber data from them.

Create a timer that runs the rsync command to pull from the active
stewards host. To have it available add it to common Hiera.

Also specify --address and the right IP to use as source IP because
lists1001.wikimedia.org also has lists.wikimedia.org on the same
interface and if we don't we run into firewall issues.

Put this in the new automation class which can be reused if we
have more use cases for automating subscriptions of certain lists.

Bug: T351202
Change-Id: I7cefee2f1bb7a964e933feabf3cb8b038c03e07a