lists/stewards: add timer to run mailman syncmembers for stewards-l
Previously we have added an rsyncd to the stewards machine and a timer
on lists1001 to sync stewards-l subscriber data from there.

Now this is a second timer to run the "mailman-wrapper syncmembers"
command to actually sync list members with the imported data.

Note that at first there is the "-n" in there that means the
command won't actually be executed. We log what it _would_ do,
as in either removing or adding list subscribers.

Later we would remove the -n and either remove logging or make
sure it gets logrotated.

Also this would be done the same way for multiple lists.

Bug: T351202
Change-Id: I2e5ecaaf0e52e8a1a1387a865717aa037b43197d