P:simplelamp2: set missing memcached_user
This was missed from 3448324288ce90e6257c221e037c13d6b15bc7b7. Set it to
the old default of 'nobody' at least for now, if/when this should be
move to 'memcached' (T273950) we can add a hiera key or just change the
hardcoded value there, but right now I just want to unbreak Puppet runs.

Change-Id: I6cad187982bdc31d979b62b9dc333c52d8a952e1