Add radosgw services to the cephosd servers
This commit causes the Ceph Rados Gateway (radosgw) to be added to the
cephosd servers. It adds a set of client capabilities for client.radosgw
and causes the relevant package to be installed.

In addition, it starts to configure /etc/ceph/ceph.conf on these
servers.

We need to be careful not to make any substantive changes to any of the
WMCS ceph clusters.

Hosts: O:ceph::server,O:wmcs::ceph::mon,O:wmcs::ceph::osd
Hosts: O:wmcs::openstack::eqiad1::control
Bug: T330152
Change-Id: I596f43d474d5d240d1ac0b50d556876e3656177f