Add a new partman reuse recipe for stat1008
stat1008 has been installed with 8 drives in a software RAID
configuration.

We now need to reimage it, but we wish to retain the /srv volume.

In order to do so we need a new reuse recipe for partman that supports
this partition format. We also use the reuse-parts-test recipe so that
we can check in the installer that the /srv volume is not going to be
overrwritten.

Bug: T329360
Change-Id: I7ef70e6b68d7c3b9522638b0c65d8e573ce9d7b1