tlsproxy::localssl: Remove support for OCSP handling
This code was added when tlsproxy was serving as our CDN TLS terminator.
For our remaining use cases of this class (TLS termination for maps
and Elastic) this isn't used and needed.

In fact, the code is alredy broken: If enabled it includes tlsproxy::ocsp,
but that class got removed in 2021 already.

Hosts: C:profile::elasticsearch::cirrus
Hosts: C:profile::maps::tlsproxy
Change-Id: I2955e20319ee39a4d74d867b407737d8e4dacfe2