P:cumin: add support for aliasing LVS host classes
This commit adds cumin aliases for the various LVS classes such as
high-traffic1, high-traffic2, etc. This allows us to do things like:
cumin "A:lvs-high-traffic1" to select all the high-traffic1 hosts in all
DCs, which is currently missing from the aliases.

When this change is merged, sre.loadbalancer.restart-pybal coobook,
which is currently broken as it depends on these aliases can be used
again to do automatic PyBal restarts.

Hosts: O:cluster::management
Change-Id: Ibc5c3243bb16651afc19c8d07f107382432e9dd3