Deprecate system::role for initial set of WMCS roles
system::role is a bit of a hack and being removed. In
21fa33b455bfd956f45525b7bb3c22c77ff3787d support was added to move the
functionality into profile::base::production and simply set the role
description via a Hiera flag, but given that none of these roles
currently set a description it can simply be removed entirely.

The role name is printed by the current motd code anyway.

Change-Id: I36adf4eff2f43a816dc32cbdfc7bae934fcac794