gitlab-settings: add timer for configure-projects
This changes ownership of /srv/gitlab-settings and runs timers for both
the LDAP sync bot and the configure-projects script under the same
(new) gitlab-settings user.

It seems unnecessary to define a new unix user for everything we might
want to run out of the gitlab-settings repo.  Currently it just does the
LDAP sync, but this patch aims to add the configure-projects script, and
I'm not clear on whether the main settings script itself is run
automatically, but maybe it should be.

This change depends on tweaking configure-projects to add a config file:

https://gitlab.wikimedia.org/repos/releng/gitlab-settings/-/merge_requests/62/d

And on a new bot account:

  - https://gitlab.wikimedia.org/configure-projects-bot
  - https://gitlab.wikimedia.org/admin/users/configure-projects-bot/edit

...for which a token has been created.

Bug: T355097
Hosts: gitlab1003.wikimedia.org,gitlab1004.wikimedia.org,gitlab2002.wikimedia.org
Change-Id: I92ad6d5d6dd08b58bee2773735ae7390f46d7c12