[beta] Add rewrite rule for sso.wikimedia.beta.wmflabs.org
It will have an extra path segment added to identify the wiki
where the user is logging in, e.g.
    /en.wikipedia.org/en/wiki/Special:Userlogin
We strip that in the rewrite rule, so the request ends up
at w/[index|api|rest].php as usual.

We might or might not do something similar in production eventually;
for now this is just to explore how reliably such a setup might work.

Soft-depends on I4a120d3024f655f8a1f020d980373ebd56b423bc.

Bug: T365162
Change-Id: Idf70fc0afb46debc32e72987da696abc23f1f65b