profile::elasticsearch::cirrus: Remove obsolete http2 parameter
HTTP2 was only ever used for the edge caches and at point accidentally
also enabled for internal services, where it's not useful per se.
As such, it was disabled in for Elastic in 2021 via these commits:
8ec00591b29af9350b1b0ab77d90c584b108757d
7c3faeaa8894324ef190915d59fa96f8c90eb654
32e818522203a2a05761fa9c53ae9cc18afd9697

As such, remove the obsolete parameter and simply pass "false"
unconditionally.

Hosts: C:profile::elasticsearch::cirrus
Change-Id: I50a13b345696e135f2e602872c64dc0695663a1f