Add wikidata history-dumps import to hdfs job
This new job is run in parallel to the existing job importing all projects
XML history onto HDFS. The reason for this specific job is to allow the other
one to finish earlier as wikidata xml-history-dumps is a lot longer to release
than other wikis dumps.
Note: for downstream jobs to be able to differenciate between all-wikis being
imported and wikidatawiki being imported, the new job uses a different success
flag.
Second note: this patch should not be deployed before refinery gets deployed with
patches:
 - https://gerrit.wikimedia.org/r/c/analytics/refinery/+/1032018
 - https://gerrit.wikimedia.org/r/c/analytics/refinery/+/1031897

Bug: T364045
Change-Id: I5e8ff7468bcf09e6d56ab05ca000f861877f14ed