benthos:cache: switch to rfc5424 format
This commit resumes the previous changes reverted, especially:

- I6836cfd828fec602c3d23e98bf38a1a05742c283
- I5398b3f134328440f35ed6df4b9d735e721f654c

To get rid of the expensive (as for CPU) grok pattern, we can switch
haproxy log format to rfc5424.

HAProxy can log to different outputs with different formats, in this
case the format for /dev/log (read by mtail) is kept the same,
removing the suffix introduced for Benthos convenience.

Logs sent explicitly to the Benthos socket instead are formatted in
rfc5424, naming all the fields with the definitive variable names used
by Benthos. This minimizes the parsing/renaming in Benthos mapping and
allow us to better reference all the various fields directly from
HAProxy configuration.

Benthos configuration has been cleaned up to use the new format and
also some now useless schema modifications has been removed too, along
with the branch processor.

The string().parse_json() workaround is copied as is from
https://github.com/benthosdev/benthos/issues/2192 and is needed due to
a bug in our production version. Can be safely removed when a newer
Benthos release is packaged and distributed.

Mtail HAProxy log parser can now be reset to the previous regex, as
it's the only one reading from the "former" log-format.

Bug: T365718
Change-Id: Idc7e5c22461f0039a77165e7bfb4c0b00855cbb3