gerrit: add parameter to toggle lfs_replica_sync
Add a parameter to allow disabling the lfs_replica_sync.

Then disable it in cloud test instances where we don't
have 2 instances.

Otherwise we get puppet errors when it's looking for
both a source and dest host but we only have one host.

And we don't need to sync anything in test instances.

Bug: T363196
Change-Id: I8d8ce13bd2a7c5d370a9ce80bbd6a70cb1399d02