gerrit: enable change.diff3ConflictView
Gerrit 3.9 introduces a diff3 based view for changes having merge
conflicts, that would also show the common base the commits are based on
which dramatically helps in resolving the conflict. It is a large
enhancement similar to `merge.conflictStyle=diff3` in cGit.

The setting can be set ahead of time on Gerrit 3.8 which will simply
ignore it.

Bug: T359821
Change-Id: I3240ff73eaba4853a628073366d01f2be631b1c2