New cephadm::rgw role
Ceph RGWs are the nodes that provide an S3 (and, optionally, swift)
api endpoint to a Ceph cluster. This is (hopefully!) the mvp for this
setup in a cephadm-based cluster.

RGW nodes are cephadm targets (which will arrange for the rgw service
to run and be available on localhost), with envoy doing the TLS
termination.

The expectation is that there will in due course be an apus cluster in
codfw too, and the RGWs will provide an S3 endpoint that replicates
between the two.

Hosts: moss-fe1002.eqiad.wmnet
Bug: T279621
Change-Id: Idf7b82e4ac5beabf959a2a478e6847035b7974d6