Temporarily disable XML dumps on snapshot hosts
We wish to defer the next XML dumps in order to give us time to
investigate T365155 and deploy a fix.

We absent the systemd timer that startes the jobs on the first of the
month.

Hosts: snapshot1010.eqiad.wmnet,snapshot1011.eqiad.wmnet
Hosts: snapshot1012.eqiad.wmnet,snapshot1013.eqiad.wmnet
Bug: T365155
Bug: T365501
Change-Id: I223e0f0e52fd77787ddc9ce692bf229453ff1251