Remove temporary firewall rule for WDQS graph_split
This temporary firewall rule was put in place in order to support the
WDQS graph split, allowing a blazegraph data transfer from stat1006 to
wdqs1022.

As stat1006 is about to be decommissioned, we want to remove this rule.

We are also working on a new cookbook that uses Transfer.py to manage
firewall rules dynamically, so the requirement for this temporary change
has been lessened.

Hosts: wdqs1022.eqiad.wmnet
Bug: T350106
Bug: T353785
Change-Id: I59f6174bc448a3e69bf80279cc885b416f08f3d0