mediawiki-image-download: Support pct based aborted runs
Why:
I wanna gauge how crucial is this functionality today. It was introduced
1.5 years ago and the environment has changed enough to warrant
gathering some data about the usefulness of it's functionality

What:
Add the ability to calculate per execution whether the script should run
or not. Hardcode the threshold to 99% at the beginning and I 'll slowly
lower it as I gauge how this affects deployment times.

Hosts: auto
Change-Id: I88e17b77bb708204f763f35499e98267909f116e