lists: Add option to switch mailman root
The new hosts are set up with a comparatively small root partition, and
a larger /srv partition. The size of /var/lib/mailman3 is larger than
the remaining space on the root partition and still growing, so we want
to move this to /srv/mailman3.

This should be mostly a no-op for files changed on the host, since no
host has the hiera option to put the data in /srv/mailman3 yet, with the
exception of the few python scripts which now include an option to
fetch the mailman base dir from the config file.

Hosts: lists1001.wikimedia.org,lists1004.wikimedia.org
Change-Id: I88808a38fa1a85a4f33277599d84c2e328647d43