deployment_server: alert on admin-ng pending changes
Now that we rely on puppet runs to generate the external services
service name -> IP mapping, we also need to make sure that we deploy
admin-ng in a timely fashion, to avoid having a sizeable diff accumulate.

This patch introduces a systemd timer per kubernetes cluster that runs
`helmfile -e $cluster diff --detailed-exitcode`, which exits with a nonzero
code when diffs are found.

We export prometheus metrics reflecting the presence/absence of a pending
diff, on which we'll be able to setup monitoring.

Bug: T331894
Hosts: deploy1002.eqiad.wmnet
Change-Id: Ie6695dc33942ca4729f78d7b64c5a27f01455f17