wmflib: add Wmflib::IP::Address::CIDR type
A new Variant type for IP address and subnet (similar to how
e.g. Stdlib::IP::Address can be either v4 or v6). Puppet stdlib v9
contains the Stdlib::IP::Address::CIDR type which we can replace this
with in due course, but it's useful to have it now (for
cephadm-related work).

Bug: T279621
Change-Id: I3981fe098ac086259e43d5b88f8fc57f3288ae3e