Configurably remove varnish handling of /beacon/event
We are decommissioning eventlogging backend components. There is one
long tail old eventlogging usage we cannot fully remove: MediaWikiPingback.

We have created a simple PHP endpoint in mediawiki.org docroot to handle
proxying /beacon/event for MediaWikiPingback to eventgate-analytics-external.

We need to remove the varnish handling of /beacon/event, so that the request
is passed to the mediawiki.org backend.

Simultaneously, we need to remove the varnishkafka-eventlogging instance
that produces requests to kafka for requests to /beacon/event.

This will ensure that only requests to mediawiki.org/beacon/event will result
in MediaWikiPingback events being produced.

Changes:

- profile::cache::kafka::eventlogging - add $ensure param.
  This can be used to ensure the varnishkafka-eventlogging instance is absent

- profile::cache::varnish::frontend - add $fe_beacon_uri_regex param.
  This can be used to alter the varnish frontend handling of /beacon/*.
  We need to disable handling of /beacon/event.

To run PCC, and to test, on cp1100 we override and set:
- profile::cache::kafka::eventlogging::ensure => absent
- profile::cache::varnish::frontend::fe_beacon_uri_regex to not match /beacon/event

Rolling this out across the frontends will be done in a future patch

Bug: T238230
Bug: T353817
Change-Id: I25716e15752773db8e93a01fe7525b1a9392a02b