helmfile: don't schedule admin-ng diff check jobs for aliases of k8s clusters
The `staging` k8s cluster is an alias to whatever cluster is the primary between
staging-eqiad and staging-codfw. We don't have a helmfile environment named `staging`,
causing the associated timer to fail.

We thus avoid creating a systemd timer operating on aliases of k8s clusters.
We also add a temporary `else` clause `absent`ing the resources that were
previously created for alias clusters, that we will remove in a subsequent patch,
once puppet has run.

Bug: T331894
Hosts: deploy2002.codfw.wmnet
Change-Id: I7b01d0295b56684f6a9a2f1fdd324899f6fac9d7