install_server: new partitioning scheme for cephadm nodes
The aim is to have a small swap partition, and the rest assigned to
the root partition (as an md-RAID device), using the two SSDs in the
chassis.

Additionally, don't wipe any LVM found, as those will be Ceph OSD
devices on storage nodes; cephadm is able to re-activate those OSDs
after a reimage.

Use this for both the frontend and backend nodes.

Bug: T279621
Change-Id: Ic457bb31ff53e0d0ecf983ec335edbf32a972595