Stop syncing swift rings on Puppet 5 frontends
This class syncs the swift rings into volatile. This is done in parallel
for Puppet 5 and Puppet 7. But at this point all Swift roles are using
Puppet 7, so we can stop the sync for the Puppet 5 masters.

After merging, these timers need to be manually disabled on
puppetmaster1001 (the class doesn't support absenting):

- fetch-rings-codfw.timer
- fetch-rings-eqiad.timer
- fetch-rings-thanos.timer

There'll also be followup patches for Thanos and media-swift to only
pass the Puppet 7 servers.

Hosts: puppetmaster1001.eqiad.wmnet, puppetmaster2001.codfw.wmnet
Bug: T365798
Change-Id: I0bcd850046f890cb5f0accc6c28bdd48794a5b65