idp: update renamed gitlab-replica OIDC service IDs
gitlab-replica was renamed to gitlab-replica-a

gitlab-replica-old was renamed to gitlab-replica-b

I already did https://gerrit.wikimedia.org/r/c/operations/puppet/+/1041750/5/hieradata/role/common/idp.yaml

but logins would not work yet.

As taavi points out there is this other place as well
but should it not be OIDC anymore?

Change-Id: Iaec6d0e57a46966a2c3db3effb042e66fcd8b18e