idp: remove gitlab from the CAS protocol section
When I edited this in I3f6ebfdf8b to reflect the new names
of gitlab-replicas I noticed I couldn't login anymore.

Logins were fixed after making the edit in the OIDC section
of this same file in Iaec6d0e57a46966a.

T320390 was the ticket to migrate from CAS protocol to OIDC
so it seems to me this part here is not used anymore and just
led to confusion.

Is that right?

Bug: T320390
Change-Id: I9d71010bcd4b3a9e54f142b111a8cbecdc37d1e7