Deprecate system::role for Kafka roles
system::role is a bit of a hack. In 21fa33b455bfd956f45525b7bb3c22c77ff3787d
support was added to move the functionality into profile::base::production
and simply set the role description via a Hiera flag.

This converts the Kafka roles. For lab-specific roles we can't annotate on the role level,
so it's simply removed for the simple broker role.

Change-Id: I6118a11151a831ca04e54d6310b4a966d21ae3c0