conftool-data: add ntp-[abc].anycast.wmnet
This is in preparation for the work under T366360, where we will be
introducing three new anycast addresses, ntp-[abc], to be announced from
all sites with the announcements being split from alternate hosts.

For more information on the idea behind the change, see the discussion
in the linked task.

The VIPs have been already been allocated on Netbox.

Bug: T366360
Change-Id: I2522483325fbf925a94b557fa249830bc3446714