durum: switch NTP peers to ntp-[abc].anycast.wmnet
To experiment with the changes in T366360, start by switching the durum
hosts to use the anycast NTP peers instead of the hardcoded Puppet ones.
We will use these hosts to test the changes before deploying it to the
wider cluster of P:systemd::timesyncd, which is 2173 hosts.

Bug: T366360
Hosts: O:durum
Change-Id: I34e64e28f42b1ae98949a541336dc4e0fb3ded38