wikikube: Use conftool for scap docker_pull_k8s
Why:
Failed nodes cause confusing errors for deployers since they are still
in the list despite being depooled.

What:
Change the way we build the dsh list from a pdb query to a conftool one.
This has the side effect of pulling in the sessionstore nodes that don't have
the mediawiki-image-download script. We override it with /bin/true on
these so we don't cause more errors.
This is a stopgap measure pending the investigation of the complete
removal of this step in T366778

Hosts: deploy1002.eqiad.wmnet,wikikube-worker1001.eqiad.wmnet,kubernetes1005.eqiad.wmnet
Bug: T367862
Change-Id: Ib070bd80e729f050a6a9452ece966b4d0778aa76